requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-retry==2.8.3
aiosignal==1.3.2
alembic==1.13.1
annotated-types==0.7.0
attrs==23.2.0
bidict==0.23.1
blinker==1.7.0
cachetools==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
dnspython==2.6.1
docstring_parser==0.16
dominate==2.9.1
email_validator==2.1.2
fake-useragent==1.5.1
Flask==2.3.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.5.0
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-aiplatform==1.61.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.65.4
h11==0.14.0
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.3.0
mailinator-python-client-2==0.0.6
Mako==1.3.5
MarkupSafe==2.1.5
mss==9.0.1
multidict==6.1.0
nodriver==0.27rc3
numpy==2.0.1
outcome==1.3.0.post0
packaging==24.1
pillow==11.0.0
playwright==1.47.0
propcache==0.2.1
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyee==12.0.0
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.2
pytz==2024.2
requests==2.31.0
rsa==4.9
selenium==4.19.0
shapely==2.0.5
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.27
streamlink==6.9.0
trio==0.25.0
trio-websocket==0.11.1
twilio==9.4.1
typing_extensions==4.10.0
undetected-chromedriver==3.5.5
urllib3==2.2.1
visitor==0.1.3
waitress==3.0.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==2.3.0
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
yarl==1.18.3
